requirements.txt
scipy
sharedmem
progressbar2
pandas
munch
torch
torchvision
matplotlib
numpy
